break out all role utils
