+++ src/shared/utils/roles/am-admin.ts
@@ -1,0 +1,5 @@
+import { UserService } from "../../services";
+
+export function amAdmin(myUserInfo = UserService.Instance.myUserInfo): boolean {
+  return myUserInfo?.local_user_view.person.admin ?? false;
+}
