+++ src/shared/utils/roles/is-mod.ts
@@ -1,0 +1,8 @@
+import { CommunityModeratorView } from "lemmy-js-client";
+
+export function isMod(
+  creatorId: number,
+  mods?: CommunityModeratorView[]
+): boolean {
+  return mods?.map(m => m.moderator.id).includes(creatorId) ?? false;
+}
